(Plan) Force re-login after certain time
Currently, we only implement timeout-relogin at client side. This is not safe. For example, Safari on iPhone won't log user out once it's run at the background. So it's necessary to let the server destroy session after certain time. To do this, we might need to implement a function called checkpw(), and call it everytime when password check is needed. Then we can add $_SESSION['createtime'] and check if session expires in this function. Also, it better to use timestamp in JS
to record the time instead of counting down.