Change encryption from GCM to CBC
GCM requires strict unique IV to ensure security. And in password
manager, we don't need authentication provided by GCM. Thus, we
switch to CBC which only requires IV to be random but do not provide
authentication. The padding attack of CBC is not related here
as attacker can't call the encryptor to encrypt his data.